[TASK] Rename namespace \\Trait\\ to \\Traits\\, due of consistency with 11.6.x
The \\Trait\\  namespace is allowed on PHP 8.0+ only.
To avoid confusions between EXT:solr 11.6 and 12.0 the namespace is adjusted to 11.6.x state.

Relates: #3843, #3844
