+++ Classes/Traits/SkipMonitoringTrait.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
  * The TYPO3 project - inspiring people to share!
  */
 
-namespace ApacheSolrForTypo3\Solr\Trait;
+namespace ApacheSolrForTypo3\Solr\Traits;
 
 use ApacheSolrForTypo3\Solr\System\Configuration\ExtensionConfiguration;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
